Rework Image notification to separate added and deleted events.
It is more efficient and interesting to have separate events
for added images and deleted images. This will also allow for
more advanced functionality in the future.

Leave the CameraControl base class to sort out the handling
of the generic image events. It dispatches the results to the
added and deleted events that are used elsewhere.