Step tethered number after end of capture.
By stepping after the capture is complete, we handle the
special case that the camera makes multiple images of
different types. It is convenient to make the images have
the file name and different suffix, so that we can sort
out that the images are different formats of the same
capture.

Also limit the number of attempts to find a free file name
for tethered capture. If nothing turns up, then there
may be a problem with the name and we don't want to get
stuck in an infinite loop.